platform--validator-workflow/requirements.txt
asgiref==3.6.0
asyncio==3.4.3
attrs==22.2.0
azure-core==1.26.3
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dnspython==2.3.0
Faker==18.4.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kafka-python==2.0.2
MarkupSafe==2.1.2
pkgutil_resolve_name==1.3.10
pycparser==2.21
pymongo==4.3.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
zipp==3.15.0